move mktiles to scripts/
