+++ scripts/mktiles/geo/load.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"os"
 
-	"github.com/ONSdigital/dp-census-atlas/mktiles/types"
+	"github.com/ONSdigital/dp-census-atlas/scripts/mktiles/types"
 
 	"github.com/twpayne/go-geom/encoding/geojson"
 )
