+++ scripts/mktiles/grid/grid.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/ONSdigital/dp-census-atlas/mktiles/types"
+	"github.com/ONSdigital/dp-census-atlas/scripts/mktiles/types"
 
 	"github.com/twpayne/go-geom"
 )
