+++ scripts/mktiles/metric/cat_test.go
@@ -3,7 +3,7 @@ package metric
 import (
 	"testing"
 
-	"github.com/ONSdigital/dp-census-atlas/mktiles/types"
+	"github.com/ONSdigital/dp-census-atlas/scripts/mktiles/types"
 )
 
 func Test_GuessTotalsCat_Errors(t *testing.T) {
