+++ scripts/mktiles/metric/ratio.go
